Cleanup:   * Remove several conversions from raw types to Integer, Boolean, etc - these are     unnecessary with Java 5 autoboxing.  Most identified via PMD.
git-svn-id: https://jamwiki.svn.sourceforge.net/svnroot/jamwiki/wiki/trunk@2525 f65b7e76-0091-dd46-87ae-4bc0c949c03e

+++ jamwiki-core/src/main/java/org/jamwiki/parser/jflex/MagicWordUtil.java
@@ -443,7 +443,7 @@ public class MagicWordUtil {
 		Date revision = null;
 		// null check needed for the test data handler, which does not implement topic versions
 		if (topic != null && topic.getCurrentVersionId() != null) {
-			topicVersion = WikiBase.getDataHandler().lookupTopicVersion(topic.getCurrentVersionId().intValue());
+			topicVersion = WikiBase.getDataHandler().lookupTopicVersion(topic.getCurrentVersionId());
 			revision = topicVersion.getEditDate();
 		}
 		formatter.setTimeZone(utc);
